Embed JS settings in base template
Embedding the global config settings into the base template allows us to
skip making an additional HTTP request on every page load.
